fix potential crash on upstream close
when doing SRT->TCP and TCP upstream closes the conn,
ngx_srt_conn_finalize is called, and it closes the upstream connection.
if a read event is received until the close event reaches the srt
thread, it may try to send to the already closed upstream connection.
setting connected to false, makes ngx_srt_proxy_srt_handler return
early, and prevents the crash.